+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Allow Stringable in array commands
 * Fixed custom provideInput() ticker
+* Use /dev/null for background manifold
 
 ## v0.11.3 (2022-12-01)
 * Fixed DataReceiver empty write error check